requirements.txt
Django==4.2.7
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

# API
djangorestframework==3.14.0

# Utilities
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0

# Vercel deployment
vercel==0.2.1

